bit more bootstrap code
