+++ config/development/sasl.conf
@@ -1,0 +1,7 @@
+  
+KafkaServer {
+   org.apache.kafka.common.security.plain.PlainLoginModule required
+   username="admin"
+   password="admin-secret"
+   user_admin="admin-secret";
+};
