+++ docker-compose.yml
@@ -1,0 +1,39 @@
+version: '3'
+services:
+  opa-firewall:
+    image: cosmtrek/air
+    environment:
+      DD_AGENT_HOST: statsd
+    volumes:
+      - ${PWD}/:/go/src/github.com/cainelli/opa-firewall
+    working_dir: /go/src/github.com/cainelli/opa-firewall/
+    command: "-c ./config/development/air.conf"
+  zookeeper:
+    image: confluentinc/cp-zookeeper:latest
+    environment:
+      ZOOKEEPER_CLIENT_PORT: 32181
+      ZOOKEEPER_TICK_TIME: 2000
+    extra_hosts:
+      - "moby:127.0.0.1"
+
+  kafka:
+    image: confluentinc/cp-kafka:latest
+    ports:
+      - 9092:9092
+      - 9091:9091
+    depends_on:
+      - zookeeper
+    volumes:
+      - ${PWD}/config/development/sasl.conf:/etc/kafka/sasl.conf
+    environment:
+      KAFKA_BROKER_ID: 1
+      KAFKA_ZOOKEEPER_CONNECT: zookeeper:32181
+      KAFKA_ADVERTISED_LISTENERS: SASL_PLAINTEXT://kafka:9092, PLAINTEXT://kafka:9091
+      KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
+      KAFKA_SASL_ENABLED_MECHANISMS: PLAIN
+      KAFKA_OPTS: -Djava.security.auth.login.config=/etc/kafka/sasl.conf
+      KAFKA_SASL_MECHANISM_INTER_BROKER_PROTOCOL: PLAIN
+      KAFKA_INTER_BROKER_LISTENER_NAME: SASL_PLAINTEXT
+      ZOOKEEPER_SASL_ENABLED: "false"
+    extra_hosts:
+      - "moby:127.0.0.1"
