+++ go.mod
@@ -3,6 +3,9 @@ module github.com/cainelli/opa-firewall
 go 1.13
 
 require (
+	github.com/Pallinder/go-randomdata v1.2.0
 	github.com/ghodss/yaml v0.0.0-20180820084758-c7ce16629ff4
+	github.com/gophercloud/gophercloud v0.8.0 // indirect
 	github.com/open-policy-agent/opa v0.17.2
+	github.com/sirupsen/logrus v1.4.1
 )
